requirements.txt
askfm-api==0.1.2 \
    --hash=sha256:5b90fb128fcc02bb54ab02be31524174ce40a3e6427e4a22c53e53ee3f17a2f6 \
    --hash=sha256:22a7321c37873fac52574e982bee6f5d2b9c9e3609c170e18b987f342c2db9b4
certifi==2020.12.5 \
    --hash=sha256:719a74fb9e33b9bd44cc7f3a8d94bc35e4049deebe19ba7d8e108280cfd59830 \
    --hash=sha256:1a4995114262bffbc2413b159f2a1a480c969de6e6eb13ee966d470af86af59c
chardet==4.0.0 \
    --hash=sha256:f864054d66fd9118f2e67044ac8981a54775ec5b67aed0441892edb553d21da5 \
    --hash=sha256:0d6f53a15db4120f2b08c94f11e7d93d2c911ee118b6b30a04ec3ee8310179fa
configly-python==1.0.8 \
    --hash=sha256:fbcfc6fec4f1081bf352e36b8bfb25cc309c008de2001a6aac423686eb673696 \
    --hash=sha256:5ff76692968f325687f86e9cbcb0503510ebd40b2d5c3483eedf32c7b4a80cfa
idna==2.10 \
    --hash=sha256:b97d804b1e9b523befed77c48dacec60e6dcb0b5391d57af6a65a312a90648c0 \
    --hash=sha256:b307872f855b18632ce0c21c5e45be78c0ea7ae4c15c828c20788b26921eb3f6
requests==2.25.1 \
    --hash=sha256:c210084e36a42ae6b9219e00e48287def368a26d03a048ddad7bfee44f75871e \
    --hash=sha256:27973dd4a904a4f13b263a19c866c13b92a39ed1c964655f025f3f8d3d75b804
urllib3==1.26.3 \
    --hash=sha256:1b465e494e3e0d8939b50680403e3aedaa2bc434b7d5af64dfd3c958d7f5ae80 \
    --hash=sha256:de3eedaad74a2683334e282005cd8d7f22f4d55fa690a2a1020a416cb0a47e73
